all operations operate by default on uint8 inputs/outputs. This prevents unnecessary conversions and reduces mem copies

+++ cuImageOps/operations/gaussianblur/gaussianblur.py
@@ -99,15 +99,13 @@ class GaussianBlur(ImageOperation):
             DataContainer: Result of the operation. Can be transfered to cpu using cpu()
         """
 
-        image_input = image_input.astype(np.float32)
-
         input_shape = image_input.shape
 
         if len(image_input.shape) <= 2:
             input_shape = (*input_shape, 1)
 
         self.input = image_input
-        self.output = np.zeros_like(self.input, dtype=np.float32)
+        self.output = np.zeros_like(self.input)
         self.intermediate_output = np.zeros_like(
             self.input, dtype=np.float32
         )  # For separable filter
